requirements.txt
aiohttp==3.7.4.post0
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.0
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.4.0
black==22.1.0
braceexpand==0.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
discord-ext-menus==1.1
discord.py @ git+https://github.com/Rapptz/discord.py@10af1ea25523a898d03eb45a0a2bb4cdbcc41bdd
docopt==0.6.2
fastapi==0.75.0
h11==0.13.0
humanfriendly==10.0
idna==3.3
import-expression==1.1.4
isort==5.10.1
jishaku==2.3.2
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pip-api==0.0.29
pipdeptree==2.2.1
pipreqs==0.4.11
platformdirs==2.5.1
pycparser==2.21
pydantic==1.9.0
PyNaCl==1.5.0
pyparsing==3.0.8
pyreadline3==3.4.1
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.2.2
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.6
wrapt==1.14.0
yarg==0.1.9
yarl==1.7.2
